Refactor entity manager code in geo_json_events integration (#89847)
* moved entity manager

* fix circular reference

* simplify new entity signal
